+++ include/mng/vector2.h
@@ -6,20 +6,20 @@ extern "C" {
 #endif
 
 typedef struct Vector2 {
-    float x;
-    float y;
+    double x;
+    double y;
 } Vector2;
 
 
 Vector2 vector2_add(Vector2 a, Vector2 b);
 Vector2 vector2_subtract(Vector2 a, Vector2 b);
-Vector2 vector2_multiply(Vector2 a, float scalar);
-float vector2_length(Vector2 a);
-float vector2_length_squared(Vector2 a);
+Vector2 vector2_multiply(Vector2 a, double scalar);
+double vector2_length(Vector2 a);
+double vector2_length_squared(Vector2 a);
 Vector2 vector2_normalized(Vector2 a);
-float vector2_dot(Vector2 a, Vector2 b);
-Vector2 vector2_rotated(Vector2 a, float rad);
-float vector2_angle(Vector2 a);
+double vector2_dot(Vector2 a, Vector2 b);
+Vector2 vector2_rotated(Vector2 a, double rad);
+double vector2_angle(Vector2 a);
 
 #ifdef __cplusplus
 }
